rename util js files and get ready for command help feature migration
